+++ test_helper.exs
@@ -1,3 +1,0 @@
-ExUnit.start()
-Ecto.Adapters.SQL.Sandbox.mode(TestWeb.Repo, :manual)
-Code.put_compiler_option(:warnings_as_errors, true)