+++ test_web/views/layout_view_test.exs
@@ -1,3 +1,0 @@
-defmodule TestWeb.LayoutViewTest do
-  use TestWeb.ConnCase, async: true
-end