+++ test_web/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule TestWeb.PageViewTest do
-  use TestWeb.ConnCase, async: true
-end