+++ lib/test_web/views/user_view.ex
@@ -1,3 +1,4 @@
 defmodule TestWeb.UserView do
+  @moduledoc false
   use TestWeb, :view
 end